move test to p2p
